Added stereo_calibration.py
This is the slightly modified varsion of a8_*.py script from BATFAST.
They had a bug in the script where it would do incomplete calibration as
it was exiting whenever a single image failed to provide corners. Fixed
that. However, the main reason for the changes was to add code to write
output intrinsics.yml and extrinsics.yml that can be used by
depth_finder.py.